Disable PG gate by default

+++ tools/container/unittests/start.sh
@@ -65,22 +65,24 @@ saladier-dbsync --config-file /tmp/saladier.conf upgrade
 run_tox MySQL runonly_pythontest_formysql
 
 ### PostgresSQL
-echo -e "${MAGENTA}Running tests under PostgresSQL${NC}"
-export SALADIER_DATABASE_TEST_CONNECTION="postgres://postgres:${SALADIER_DB_PASSWORD}@${POSTGRES_PORT_5432_TCP_ADDR}/saladier"
+if [[ -n ${POSTGRES} ]];then
+        echo -e "${MAGENTA}Running tests under PostgresSQL${NC}"
+    export SALADIER_DATABASE_TEST_CONNECTION="postgres://postgres:${SALADIER_DB_PASSWORD}@${POSTGRES_PORT_5432_TCP_ADDR}/saladier"
 
-PGPASSWORD=${SALADIER_DB_PASSWORD} dropdb --if-exists -U postgres -h${POSTGRES_PORT_5432_TCP_ADDR} saladier
-PGPASSWORD=${SALADIER_DB_PASSWORD} createdb -U postgres -h${POSTGRES_PORT_5432_TCP_ADDR} saladier
+    PGPASSWORD=${SALADIER_DB_PASSWORD} dropdb --if-exists -U postgres -h${POSTGRES_PORT_5432_TCP_ADDR} saladier
+    PGPASSWORD=${SALADIER_DB_PASSWORD} createdb -U postgres -h${POSTGRES_PORT_5432_TCP_ADDR} saladier
 
-cat <<EOF>/tmp/saladier.conf
+    cat <<EOF>/tmp/saladier.conf
 [database]
 connection=${SALADIER_DATABASE_TEST_CONNECTION}
 EOF
 
-# Create schema and upgrade
-saladier-dbsync --config-file /tmp/saladier.conf create_schema
-saladier-dbsync --config-file /tmp/saladier.conf upgrade
+    # Create schema and upgrade
+    saladier-dbsync --config-file /tmp/saladier.conf create_schema
+    saladier-dbsync --config-file /tmp/saladier.conf upgrade
 
-run_tox PostgresSQL runonly_pythontest_formysql
+    run_tox PostgresSQL runonly_pythontest_formysql
+fi
 
 ### SQLITE
 # NOTE(chmou): need to be at the end after the others cause it generates the
